renamed role_group_role to group_membership moved util functions around to avoid circular references
