+++ moebot_bot/util/util.go
@@ -9,16 +9,10 @@ package util
 import (
 	"database/sql"
 	"fmt"
-	"log"
 	"regexp"
 	"strconv"
 	"strings"
 	"sync"
-
-	"github.com/camd67/moebot/moebot_bot/util/db"
-	"github.com/camd67/moebot/moebot_bot/util/moeDiscord"
-
-	"github.com/bwmarrin/discordgo"
 )
 
 const (
@@ -34,57 +28,6 @@ type SyncUIDByChannelMap struct {
 type SyncCooldownMap struct {
 	sync.RWMutex
 	M map[string]int64
-}
-
-func IntContains(s []int, e int) bool {
-	return IntIndexOf(s, e) > -1
-}
-
-func IntRemove(s []int, e int) []int {
-	index := IntIndexOf(s, e)
-	if index > -1 {
-		s = append(s[:index], s[index+1:]...)
-	}
-	return s
-}
-
-func IntIndexOf(s []int, e int) int {
-	for i, a := range s {
-		if a == e {
-			return i
-		}
-	}
-	return -1
-}
-
-func StrContains(s []string, e string, caseInsensitive int) bool {
-	for _, a := range s {
-		if caseInsensitive == CaseInsensitive {
-			if strings.EqualFold(e, a) {
-				return true
-			}
-		} else {
-			if a == e {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func StrContainsPrefix(s []string, e string, caseInsensitive int) bool {
-	for _, a := range s {
-		if caseInsensitive == CaseInsensitive {
-			if strings.HasPrefix(strings.ToUpper(a), strings.ToUpper(e)) {
-				return true
-			}
-		} else {
-			if strings.HasPrefix(a, e) {
-				return true
-			}
-		}
-	}
-	return false
 }
 
 func MakeAlphaOnly(s string) string {
@@ -189,82 +132,4 @@ func ParseIntervalToISO(interval string) (string, error) {
 	}
 	intervalString := strings.Trim(b.String(), "T")
 	return intervalString, nil
-}
-
-func RetrieveBasePermissions(session *discordgo.Session, channel *discordgo.Channel, role *discordgo.Role, flags []int) map[int]bool {
-	result := make(map[int]bool)
-	permission, ok := moeDiscord.FindPermissionByRoleID(channel.PermissionOverwrites, role.ID)
-	if ok {
-		mapPermissions(result, permission, flags)
-	}
-	if !ok || unsetFlags(permission, flags) { //no overwrite defined for the channel, looking in parent category
-		parent, _ := session.Channel(channel.ParentID)
-		permission, ok = moeDiscord.FindPermissionByRoleID(parent.PermissionOverwrites, role.ID)
-		if ok {
-			mapPermissions(result, permission, flags)
-		}
-		if !ok || unsetFlags(permission, flags) { //no overwrite defined for the channel, using role permissions
-			permission = &discordgo.PermissionOverwrite{
-				ID:   role.ID,
-				Type: "role",
-			}
-			for _, f := range flags {
-				if role.Permissions&f != 0 {
-					permission.Allow = permission.Allow | f
-				} else {
-					permission.Deny = permission.Deny | f
-				}
-			}
-			mapPermissions(result, permission, flags)
-		}
-	}
-	return result
-}
-
-func GetEveryoneRoleForServer(session *discordgo.Session, serverID int) *discordgo.Role {
-	server, err := db.ServerQueryById(serverID)
-	if err != nil {
-		log.Println(fmt.Sprintf("Failed to retrieve server informations for Server ID: %v. ", serverID), err)
-		return nil
-	}
-	roles, err := session.GuildRoles(server.GuildUid)
-	if err != nil {
-		log.Println("Failed to retrieve roles informations for Guild UID: "+server.GuildUid+". ", err)
-		return nil
-	}
-	return moeDiscord.FindRoleByName(roles, "@everyone")
-}
-
-func GetCurrentRolePermissionsForChannel(session *discordgo.Session, channelUID string, roleUID string) (*discordgo.PermissionOverwrite, error) {
-	channel, err := session.Channel(channelUID)
-	if err != nil {
-		return nil, err
-	}
-	if p, ok := moeDiscord.FindPermissionByRoleID(channel.PermissionOverwrites, roleUID); !ok {
-		return &discordgo.PermissionOverwrite{
-			ID:   roleUID,
-			Type: "role",
-		}, nil
-	} else {
-		return p, nil
-	}
-}
-
-func unsetFlags(permission *discordgo.PermissionOverwrite, flags []int) bool {
-	for _, f := range flags {
-		if permission.Allow&f == 0 && permission.Deny&f == 0 {
-			return true
-		}
-	}
-	return false
-}
-
-func mapPermissions(base map[int]bool, permission *discordgo.PermissionOverwrite, flags []int) {
-	for _, f := range flags {
-		if _, ok := base[f]; !ok { //only do this if the flag is unset, to allow hierarchy assignations
-			if permission.Allow&f != 0 || permission.Deny&f != 0 {
-				base[f] = permission.Allow&f != 0
-			}
-		}
-	}
 }
